Logi institute and other details
refs #722
I've added the logo on the welcome pages (en|fr|nl). The logo of the institute
on the about page is now located above the partner's title, as requested in
the issue.
I've made an image map for the top menu background image, this way people can
click on DaRWIN or on the Drop-image and return to the Welcome page, behaviour
people are most likely used to (I know I am >.<)